scripts/models/requirements.txt
pandas
numpy
scikit-learn
matplotlib
scikit-image
pymongo
python-dotenv